+++ src/usecases/page/FindPageList.ts
@@ -1,5 +1,4 @@
-import { Page, PageStatus } from '../../domains/Page';
-import { User } from '../../domains/User';
+import { Page } from '../../domains/Page';
 import { PaginationQuery, PaginationOptions } from '../../interfaces/pagination';
 import { IPageRepository } from '../../repositories/IPageRepository';
 
@@ -10,27 +9,7 @@ export class FindPageList {
     this.pageRepository = pageRepository;
   }
 
-  execute(user: User, status: PageStatus, isFavorite: boolean, sort: string, page: number, limit: number): Promise<Page> {
-    const query: PaginationQuery = {
-      createdUser: user._id,
-      status,
-    };
-
-    if (isFavorite != null) {
-      query.isFavorite = isFavorite;
-    }
-
-    const options: PaginationOptions = {
-      page,
-      limit,
-    };
-
-    if (sort != null) {
-      const sortOrder = sort.startsWith('-') ? -1 : 1;
-      const sortKey = sortOrder === -1 ? sort.slice(1) : sort;
-      options.sort = { [sortKey]: sortOrder };
-    }
-
+  execute(query: PaginationQuery, options: PaginationOptions): Promise<Page> {
     return this.pageRepository.findPageList(query, options);
   }
 }
